docker/process/requirements.txt
attrdict==2.0.1
numpy==1.19.2
opencv_python==4.5.5.62
Pillow==9.2.0
PyYAML==6.0
torch==1.9.0
tqdm==4.50.2